requirements.txt
pandas>=2.0
pyarrow>=11.0
pydantic>=1.10
SQLAlchemy>=1.4,<2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
apache-airflow==2.7.3
pytest>=7.0

# black==24.8.0
# ruff==0.4.8
